initial pattern abstraction and move matching to src
