+++ src/ast/euf/CMakeLists.txt
@@ -10,6 +10,7 @@ z3_add_component(euf
     euf_mam.cpp
     euf_plugin.cpp
     euf_specrel_plugin.cpp
+    ho_matcher.cpp
   COMPONENT_DEPENDENCIES
     ast
     util
